extractor-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
python-multipart==0.0.9
google-genai==1.0.0
supabase==2.9.0
stripe==10.0.0
